Change go dependency codegen
